Add files via upload
You will build a classification model that guesses whether a movie is a comedy or a thriller by using only the number of times chosen words appear in the movie's screenplay. By the end of the project, you should know how to:

Build a k-nearest-neighbors classifier.
Test a classifier on data.